Add explicit type conversions where necessary: components/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2947886
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#890878}

+++ components/spellcheck/browser/spellcheck_host_metrics_unittest.cc
@@ -122,8 +122,8 @@ TEST_F(SpellcheckHostMetricsTest, RecordAcceptLanguageStats) {
 
   for (size_t i = 0; i < base::size(histogram_names); ++i) {
     histogram_tester.ExpectTotalCount(histogram_names[i], 1);
-    histogram_tester.ExpectBucketCount(histogram_names[i], expected_counts[i],
-                                       1);
+    histogram_tester.ExpectBucketCount(histogram_names[i],
+                                       static_cast<int>(expected_counts[i]), 1);
   }
 }
 
@@ -140,8 +140,8 @@ TEST_F(SpellcheckHostMetricsTest, RecordSpellcheckLanguageStats) {
 
   for (size_t i = 0; i < base::size(histogram_names); ++i) {
     histogram_tester.ExpectTotalCount(histogram_names[i], 1);
-    histogram_tester.ExpectBucketCount(histogram_names[i], expected_counts[i],
-                                       1);
+    histogram_tester.ExpectBucketCount(histogram_names[i],
+                                       static_cast<int>(expected_counts[i]), 1);
   }
 }
 #endif  // defined(OS_WIN)
